fix router bug
Fixed: if user tried to sign in before email verif, the user has to be signed in briefly to resend email verif, this caused the router to try to redirect as login shouldn't be accessible when signed in, this prevented the user from being signed out properly and broke the nav bar. Added an extra check on the router to only redirect when signed in if the user is verified.